Calculate timings for each trace
This commit updates the span parsing logic to include a computation of the
inclusive and exclusive timings for each span in each trace. The inclusive
time for a span is the total time in milliseconds spent on the span. The
exclusive time is equal to the inclusive time minus the inclusive time of
any child spans.

Closes #1.