Add UUID to each Trace
This commit adds a random UUIDv4 to each Trace that is parsed from a log. This
id is set on the root entry of the trace and propagates to all child entries in
the trace. Parsing the same log file twice will yield different trace ids as
the assignment is random.