Imbue Span class with OpenTracing data model
This commit adds some items from the [OpenTracing data model][data-model]
to the Span class:

  - start_time, finish_time: Start and finish time of the operation.
  - context: Contains the Trace and Span ids.
  - tags: Hash of data associated with the span.
  - references: Array of references to other spans.

Notably, this commit does not implement support for the `logs` attribute
described by the model as collecting logs is currently out of scope for
the profile parser.

The Trace class is updated to fill in some of this information as part of
finalization.

  [data-model]: https://github.com/opentracing/specification/blob/1.1/specification.md#the-opentracing-data-model