Parse request id from PROFILE lines
The Puppet profiler includes an ID in each log line that is assigned uniquely
assigned for each HTTP request. This ID can be used to separate logs from
different requests in the case of Rack-based Puppet Masters instead of
relying on the Java thread id which only exists in Puppet Server.

This change also causes the LogParser to consume the leading "PROFILE" string
from each log message which will simplify parsing down in the TraceParser.