Add compatibility with Ruby 2.0
This commit updates the Puppet Profile Parser to be compatible with Ruby 2.0.
This is the Ruby version bundled with the CentOS 7 base repositories and is
therefore likely to be the first hit on the $PATH for many users. Achieving
compatibility was a simple matter of swapping the `.to_h` method used to
convert from Array to hash for the `Hash[ ]` constructor.

Ref. #12.