Extract exit logic from CLI class
This commit updates the CLI class to remove any `exit` calls. The `run` method
of CLI instances now returns an integer that can be set as an exit code. This
allows all parts of the PuppetProfileParser module to be called as a library
without causing termination of the calling program.