Update Ruby versions in Travis matrix
This commit replaces the Ruby 2.1.9 test with a test against Ruby 2.5.7
and updates the Ruby 2.4 test to 2.4.7. These versions correspond to
what is currently used by Puppet 5.5.x and 6.x and drops Puppet 4.10.x
which is end of life.

The test against Ruby 2.0 is retained in order to maintain backwards
compatibility with older versions.